wip: Auto generate dark mode icon variants
I'm not super happy with this, as it has some artifacting, doesn't look
perfect, and doesn't work for all my icons.

Unfortunately, I've lost track of my original sketch file for the icons
(I do have access to a recent one but it only has the last couple
browsers I added and a hacked version of the frame) so I can't generate
proper versions of these.

Related to #6